Movement / Jump working. Camera added.
Movement and jumping is working, though jump is taken straight form unreals premade jump function, whereas movement is still handled by custom c++. Camera is added, but it fails to maintain a set distance to the character when the character is pushed around because of collision.

Project builds and compiles without errors or warnings at this stage.