Implemented Glide and StopGliding functions
Implemented Glide and StopGliding functions. The ability to glide is locked by default in beginning of game - for debugging/testing the ability can be unlocked by setting bCanGlide = true in the ABunnyCharacter.cpp file.

Made other minor alterations to unify the code, implenented acceleration as specified in the game design document.