fixed the RootComponent bug and reimplemented AddMovementInput
This leaves us with the unfortunate situation that the ground "holds on"
to the character as long as we are in walking mode so that it is now
necessary to "bump up" the character in an unnatural manner when
starting to climb. Could maybe be fixed by using flying mode when
climbing.