Added DeathMenu, HelpMenu, button funtionality to swtich between menus and Death() funtion
Added a death function to the bunny character class. Added the DeathMenu and a HelpMenu (with controls shown) and removed the Prototype HUD. It should now be possible to navigate through menus and the buttons all appear to work. It is possible to restart a level on death or from main menu (i.e. open the level you just exited from the start).

The death function, when called, will bring the player to the Death Menu where she may quit the game entirely, return to main menu or restart the level.