Tweaked glide functionality by changing a few lines in Tick()
Added functionality for interrupting upwards movement to immediately begin gliding. Attempting to press Glide button will now only not call Glide if the GetCharacterMovement->Velocity.Z is 0 when pressed. (I.e. when not falling or jumping. This will create an edge case where on the top of a jump the velocity might be 0 as the upwards force and gravity matches each other and as such fials to execute Glide().

In future I would probably wih to look into the IsFalling component and use it appropriately, but it did not work as expected so will stick with velocity.Z for now.