Added an implementation of Level 2, tweaked menus, added some comments in code and blueprints
Created level 2. It was necessary to add collision to meshes and make soem modifications to those uasset files that are not committed here. The level is therefore unlikely to work very well in an envirement where these changes are not reflected.

Added some comments in the BunnyCharacter class. Added more help to the help menu.