added all improved assets and room segments in _new folders
They are nicely sorted in meshes and materials.
Further sorted in assets (things like chairs, cups, books, furniture)
and rooms (segmented walls, floors, door).
Characters are also in a separate folder, but they are not added yet.
I have made good collision for all the new versions of the meshes.