Added sounds
The afterlife is no longer a place of death where absolute silence reigns. Now, there is some merry sounds for button clicks, jumping, gliding, climbing, scaring, getting hit, dying and getting to the victory screen.

Due to the nature of our program, some of the sounds are played through blueprint, others relating to the bunny character are played from c++ source files.

While bad form to commit and push directly to master, circumstances made it quicker and more convenient. A new executable (_c) has been made from this version and seems to be in working order.